+++ src/avatar/avatar-info.tsx
@@ -12,13 +12,11 @@ interface InfoAvatarProps {
 
 export const fontSizes: Record<Size, number> = {
   [Size.Size16]: 9,
-  [Size.Size18]: 9,
   [Size.Size20]: 9,
   [Size.Size24]: 11,
   [Size.Size28]: 12,
   [Size.Size32]: 14,
   [Size.Size40]: 16,
-  [Size.Size48]: 16,
   [Size.Size56]: 22,
 };